# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlightGo Backend API",
    "version": "1.0.0",
    "description": "FlightGo Backend API"
  },
  "servers": [
    {
      "url": "undefined",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Flight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The flight ID"
          },
          "flight_number": {
            "type": "integer",
            "description": "The route flight number"
          },
          "airline_code": {
            "type": "string",
            "description": "The airline code (2 uppercase letters)"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "The flight date (YYYY-MM-DD)"
          },
          "available_tickets": {
            "type": "integer",
            "description": "Number of available tickets"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "The price of the flight"
          },
          "version": {
            "type": "integer",
            "description": "Version for optimistic concurrency"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/flight": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new flight",
        "tags": [
          "Flight"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flight"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flight created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all flights",
        "tags": [
          "Flight"
        ],
        "responses": {
          "200": {
            "description": "List of flights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          }
        }
      }
    },
    "/flight/search": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Search flights with filters",
        "tags": [
          "Flight"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "departure_airport",
            "schema": {
              "type": "string"
            },
            "description": "Departure airport code (3 uppercase letters)"
          },
          {
            "in": "query",
            "name": "destination_airport",
            "schema": {
              "type": "string"
            },
            "description": "Destination airport code (3 uppercase letters)"
          },
          {
            "in": "query",
            "name": "min_departure_time",
            "schema": {
              "type": "string"
            },
            "description": "Minimum departure time (HH:mm)"
          },
          {
            "in": "query",
            "name": "max_departure_time",
            "schema": {
              "type": "string"
            },
            "description": "Maximum departure time (HH:mm)"
          },
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Flight date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "min_price",
            "schema": {
              "type": "number"
            },
            "description": "Minimum price"
          },
          {
            "in": "query",
            "name": "max_price",
            "schema": {
              "type": "number"
            },
            "description": "Maximum price"
          },
          {
            "in": "query",
            "name": "min_duration",
            "schema": {
              "type": "integer"
            },
            "description": "Minimum flight duration in minutes"
          },
          {
            "in": "query",
            "name": "max_duration",
            "schema": {
              "type": "integer"
            },
            "description": "Maximum flight duration in minutes"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching flights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input parameters"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          }
        }
      }
    },
    "/flight/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get an existing flight",
        "tags": [
          "Flight"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The flight ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flight found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "404": {
            "description": "Flight not found"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an existing flight",
        "tags": [
          "Flight"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The flight ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "available_tickets": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "number",
                    "format": "float"
                  },
                  "version": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flight updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Flight not found"
          },
          "409": {
            "description": "Version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "currentVersion": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a flight",
        "tags": [
          "Flight"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The flight ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Flight deleted successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Flight not found"
          }
        }
      }
    },
...
  },
...
}